Change view of keyboard shortcut:
Original: Display all shortcuts in form of `Example(X)`
Now: Display the shortcuts in English word, if any.
i.e. `E&amp;xample`

Originally it is to follow the pattern of Taiwan translation.
However, the previous commit had added a lot of shortcuts.
It is better to adapt new strategy for the UI.